c2rust refactor -r inplace --cargo remove_redundant_casts

+++ chain.rs
@@ -145,7 +145,8 @@ unsafe extern "C" fn build_lut_matrix(mut lut: *mut lutType) -> matrix {
     } else {
         memset(&mut result as *mut matrix as *mut libc::c_void,
                0i32,
-               ::std::mem::size_of::<matrix>() as usize);
+               
+               ::std::mem::size_of::<matrix>());
         result.invalid = 1i32 != 0
     }
     return result;
@@ -175,7 +176,8 @@ unsafe extern "C" fn build_mAB_matrix(mut lut: *mut lutmABType) -> matrix {
     } else {
         memset(&mut result as *mut matrix as *mut libc::c_void,
                0i32,
-               ::std::mem::size_of::<matrix>() as usize);
+               
+               ::std::mem::size_of::<matrix>());
         result.invalid = 1i32 != 0
     }
     return result;
@@ -334,7 +336,7 @@ unsafe extern "C" fn qcms_transform_module_clut_only(mut transform:
     let mut b_table: *mut f32 = (*transform).b_clut;
      let mut i:  size_t =  0u64;
     while i < length {
-        debug_assert!((*transform).grid_size as i32 >= 1i32);
+         debug_assert!((*transform).grid_size as i32 >= 1i32);
         let fresh12 = src;
         src = src.offset(1);
         let mut linear_r: f32 = *fresh12;
@@ -478,7 +480,7 @@ unsafe extern "C" fn qcms_transform_module_clut(mut transform:
     let mut b_table: *mut f32 = (*transform).b_clut;
      let mut i:  size_t =  0u64;
     while i < length {
-        debug_assert!((*transform).grid_size as i32 >= 1i32);
+         debug_assert!((*transform).grid_size as i32 >= 1i32);
         let fresh18 = src;
         src = src.offset(1);
         let mut device_r: f32 = *fresh18;
@@ -998,8 +1000,8 @@ unsafe extern "C" fn qcms_transform_module_matrix(mut transform:
 unsafe extern "C" fn qcms_modular_transform_alloc()
  -> *mut qcms_modular_transform {
     return calloc(1,
-                  ::std::mem::size_of::<qcms_modular_transform>() as
-                      usize) as *mut qcms_modular_transform;
+                  
+                  ::std::mem::size_of::<qcms_modular_transform>()) as *mut qcms_modular_transform;
 }
 unsafe extern "C" fn qcms_modular_transform_release(mut transform:
                                                         *mut qcms_modular_transform) {
@@ -1164,8 +1166,7 @@ unsafe extern "C" fn qcms_modular_transform_create_mAB(mut lut:
                     } else {
                         append_transform(transform, &mut next_transform);
                         clut_length =
-                            (::std::mem::size_of::<f32>() as
-                                 usize as f64 *
+                            (::std::mem::size_of::<f32>() as f64 *
                                  ((*lut).num_grid_points[0usize] as
                                          f64).powf(
                                      3f64) *
@@ -1329,8 +1330,7 @@ unsafe extern "C" fn qcms_modular_transform_create_lut(mut lut: *mut lutType)
             if !transform.is_null() {
                 append_transform(transform, &mut next_transform);
                 in_curve_len =
-                    (::std::mem::size_of::<f32>() as
-                         usize).wrapping_mul((*lut).num_input_table_entries
+                    (::std::mem::size_of::<f32>()).wrapping_mul((*lut).num_input_table_entries
                                                          as
                                                          usize).wrapping_mul(3);
                 in_curves = malloc(in_curve_len) as *mut f32;
@@ -1354,8 +1354,7 @@ unsafe extern "C" fn qcms_modular_transform_create_lut(mut lut: *mut lutType)
                         (*lut).num_input_table_entries;
                     // Prepare table
                     clut_length =
-                        (::std::mem::size_of::<f32>() as
-                             usize as f64 *
+                        (::std::mem::size_of::<f32>() as f64 *
                              ((*lut).num_clut_grid_points as
                                      f64).powf(
                                  3f64) *
@@ -1375,8 +1374,7 @@ unsafe extern "C" fn qcms_modular_transform_create_lut(mut lut: *mut lutType)
                             (*lut).num_clut_grid_points as uint16_t;
                         // Prepare output curves
                         out_curve_len =
-                            (::std::mem::size_of::<f32>() as
-                                 usize).wrapping_mul((*lut).num_output_table_entries
+                            (::std::mem::size_of::<f32>()).wrapping_mul((*lut).num_output_table_entries
                                                                  as
                                                                  usize).wrapping_mul(3);
                         out_curves =
@@ -1646,7 +1644,7 @@ unsafe extern "C" fn qcms_modular_transform_create_output(mut out:
             }
         }
     } else {
-        debug_assert!(false, "Unsupported output profile workflow.");
+         debug_assert!(false, "Unsupported output profile workflow.");
         return 0 as *mut qcms_modular_transform
     }
     match current_block {
@@ -1798,7 +1796,7 @@ unsafe extern "C" fn qcms_modular_transform_create(mut in_0:
                                     return first_transform
                                 }
                             } else {
-                                debug_assert!(false, "output color space not supported");
+                                 debug_assert!(false, "output color space not supported");
                             }
                         }
                     }
@@ -1806,7 +1804,7 @@ unsafe extern "C" fn qcms_modular_transform_create(mut in_0:
             }
         }
     } else {
-        debug_assert!(false, "input color space not supported");
+         debug_assert!(false, "input color space not supported");
     }
     qcms_modular_transform_release(first_transform);
     return 0 as *mut qcms_modular_transform;
@@ -1877,7 +1875,7 @@ unsafe extern "C" fn qcms_modular_transform_data(mut transform:
                                                  _: *mut f32,
                                                  _: *mut f32,
                                                  _: size_t) -> ()) {
-            debug_assert!(false, "Unsupported transform module");
+             debug_assert!(false, "Unsupported transform module");
             return 0 as *mut f32
         }
         if (*transform).grid_size as i32 <= 0i32 &&
@@ -1895,7 +1893,7 @@ unsafe extern "C" fn qcms_modular_transform_data(mut transform:
                                                       _: *mut f32,
                                                       _: *mut f32,
                                                       _: size_t) -> ())) {
-            debug_assert!(false, "Invalid transform");
+             debug_assert!(false, "Invalid transform");
             return 0 as *mut f32
         }
         (*transform).transform_module_fn.expect("non-null function pointer")(transform,
